requirements.txt
asgiref==3.7.2
Django==4.2.20
firebase-admin==6.2.0
google-auth==2.22.0
google-cloud-storage==2.10.0
pillow==11.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.6.0
gunicorn==21.2.0
django-crispy-forms==2.1
django-environ==0.11.2
django-widget-tweaks==1.5.0
